refactor: rename physics.amplitude to model.sympy
