+++ src/tensorwaves/model/sympy.py
